+++ modules/rendering/include/model/node.h
@@ -66,6 +66,20 @@ struct Node {
    */
   glm::mat4 GetGlobalMatrix();
 
+  /**
+   * @brief Applies the initial transformation to the specified node and its
+   * child nodes recursively.
+   *
+   * This function applies the initial transformation to the given node and its
+   * child nodes recursively. It ensures that the transformation is propagated
+   * down the hierarchy of nodes, allowing for hierarchical transformations in
+   * the scene graph.
+   *
+   * @param node A shared pointer to the node to which the initial
+   * transformation should be applied.
+   */
+  void InitialTransform();
+
   std::string name;
   std::shared_ptr<Mgtt::Rendering::Node> parent;
   std::shared_ptr<Mgtt::Rendering::Mesh> mesh;